Applied the changes mentioned in the PR review
